Revert "제품 상세 페이지 뼈대 구현"
