+++ lib/main.dart
@@ -1,13 +1,10 @@
 import 'package:baljachwi_project/screens/catalog_list_screen.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui';
-
-import 'package:baljachwi_project/screens/mypage/mypage_screen.dart';
 import 'package:baljachwi_project/screens/nonmember_home_screen.dart';
 import 'package:baljachwi_project/screens/catalog_list_screen.dart';
 import 'package:baljachwi_project/screens/nonmember_order_screen.dart';
 import 'package:baljachwi_project/screens/product_details_screen.dart';
-
 
 void main() {
   runApp(const MyApp());
@@ -20,7 +17,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       //debugShowCheckedModeBanner: false,
-      home: mypage(),
+      home: productDetails(),
     );
   }
 }
