--- lib/screens/mypage/frequently_asked_q&a_screen.dart
@@ -1,44 +1,0 @@
-import 'package:flutter/material.dart';
-
-class frequentQuestions extends StatelessWidget {
-  const frequentQuestions({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          '자주하는 질문',
-          style: TextStyle(
-            color: Colors.black,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        centerTitle: true,
-        leading: IconButton(
-          icon: Icon(Icons.navigate_before),
-          color: Colors.black,
-          iconSize: 30,
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
-        backgroundColor: Colors.white,
-        elevation: 0,
-      ),
-      body: Column(
-        mainAxisSize: MainAxisSize.max,
-        children: <Widget>[
-          Container(
-            height: 2,
-            color: Color(0xffc0c0c0),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/*
-1. 1:1 문의에서 연결
-*/
