--- lib/screens/mypage/order_history_screen.dart
@@ -1,46 +1,0 @@
-import 'package:flutter/material.dart';
-
-class orderHistory extends StatelessWidget {
-  const orderHistory({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          '주문 내역',
-          style: TextStyle(
-            color: Colors.black,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        centerTitle: true,
-        leading: IconButton(
-          icon: Icon(Icons.navigate_before),
-          color: Colors.black,
-          iconSize: 30,
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
-        backgroundColor: Colors.white,
-        elevation: 0,
-      ),
-      body: Column(
-        mainAxisSize: MainAxisSize.max,
-        children: <Widget>[
-          Container(
-            height: 2,
-            color: Color(0xffc0c0c0),
-          ),
-          Expanded(
-            child: Container(
-              height: double.infinity,
-              color: Color(0xfff2f2f2),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
